+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## **WORK IN PROGRESS**
 
 * (theimo1221) Preparations for the implementation of SmartGardenDevices
+* (theimo1221) Add Smart-Garden Sensor (temperature and humidity of soil)
 
 ## 3.0.0-alpha.61 (2024-05-10)